Add PartialAvailabilityPolicyFail, refactor tests to TagPartitionedLogSystemRecoveryTests.actor.cpp
